requirements.txt
python-wordpress-xmlrpc
requests
openai
googletrans==3.1.0a0